backend/requirements.txt
# --- Pakiety, które już były ---
asarPy==1.0.1
Cython==3.0.11
Mako==1.3.8
Markdown==3.5.2
MarkupSafe==3.0.2
meson==1.5.2
packaging==24.2
pycairo==1.27.0
Pygments==2.18.0
PyGObject==3.50.0
setuptools==75.6.0
six==1.17.0
wheel==0.45.1

# --- Pakiety Django/REST ---
Django==5.1.4
djangorestframework==3.14
django-cors-headers==3.14.0

# --- Pakiety do analizy NLP / ML ---
nltk==3.8.1
spacy==3.8.2
scikit-learn==1.3.2
pandas==2.2.3
numpy==1.26.4
gensim==4.3.3
